Separate Qio and remove redundancy in flux calculation.
